backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cy0802 <amyyoyo0802@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
flask = "3.0.3"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.3"
markupsafe = "^2.1.5"
werkzeug = "^3.0.2"
flask-sqlalchemy = "3.0.3"
python-dotenv = "^1.0.1"
pymysql = "1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
